+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -145,7 +145,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KAB6F73TR6YZ39DHREPD5FKV"),
+  _id: z.string().default("01KABAYG2AZV2VJH6AHBVQD02S"),
   display_name: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$inboundSchema,
@@ -188,7 +188,7 @@ export const RetrieveDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   RetrieveDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KAB6F73TR6YZ39DHREPD5FKV"),
+  id: z.string().default("01KABAYG2AZV2VJH6AHBVQD02S"),
   displayName: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$outboundSchema,