+++ packages/orq-rc/RELEASES.md
@@ -5109,3 +5109,13 @@ Based on:
 - [typescript v4.0.0-rc.44] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.44] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.44 - packages/orq-rc
+
+## 2025-11-18 11:19:13
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.658.2 (2.755.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.45] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.45 - packages/orq-rc